requirements.txt
ansi2html==1.8.0
Appium-Python-Client==2.11.1
Appium-UIAutomation==0.1.0
asgiref==3.5.2
attrs==23.1.0
bcrypt==4.1.3
behave==1.2.6
blinker==1.7.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==38.0.4
django-background-tasks==1.2.5
django-compat==1.0.15
docutils==0.20.1
ec==0.2.5
ez-setup==0.9
exceptiongroup==1.1.3
Flask==3.0.2
gherkin-official==29.0.0
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-oauth2-tool==0.0.3
googleapis-common-protos==1.62.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmlmin==0.1.12
hyperframe==6.0.1
idna==3.4
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
kaitaistruct==0.10
ldap3==2.9.1
looseversion==1.3.0
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.3
mitmproxy_wireguard==0.1.23
msgpack==1.0.8
mysql-connector-python==8.3.0
oauth2client==3.0.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
paramiko==3.4.0
parse==1.20.1
parse-type==0.6.2
passlib==1.7.4
pip==25.0.1
Pillow==10.0.1
playwright==1.37.0
pluggy==1.5.0
protobuf==4.25.2
psutil==5.9.8
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyee==9.0.4
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyperclip==1.8.2
PySocks==1.7.1
pytest==7.4.0
pytest-base-url==2.0.0
pytest-html==4.1.1
pytest-html-reporter==0.2.9
pytest-metadata==3.0.0
pytest-playwright==0.4.2
pytest-reporter==0.5.3
pytest-runner==6.0.0
pytest-variables==3.0.0
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.40
selenium==4.11.2
setuptools==72.2.0
six==1.16.0
slack_sdk==3.26.2
sniffio==1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.4
trio==0.22.2
trio-websocket==0.10.3
uiautomator==1.0.2
uritemplate==4.1.1
urllib3==2.0.4
urwid==2.1.2
webdriver-manager==4.0.0
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.15.0
wsproto==1.2.0
zipp==3.17.0
zstandard==0.19.0
gspread==6.1.4
aioquic==0.9.25
httplib2==0.22.0
pika==1.3.2
pylsqpack==0.3.18
pyparsing==3.1.1
pytest-bdd==7.2.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
service-identity==24.1.0
google-genai==1.7.0
aiohttp==3.11.16
mcp==1.6.0
google-generativeai==0.8.4
greenlet==2.0.2
typing_extensions==4.11.0
